coded buy functionality using fetch() to razorpay endpoint

+++ bon_app/models.py
@@ -5,10 +5,9 @@ from django.utils.text import slugify
 from django.db import models
 from django.contrib.auth.models import User
 
-
 class Payment(models.Model):
     order_id = models.CharField(max_length=100)
-    amount = models.DecimalField(max_digits=10, decimal_places=2)
+    amount = models.IntegerField()
     status = models.CharField(max_length=20, default="pending")
     created_at = models.DateTimeField(auto_now_add=True)
